Rename venues controller to plural form
